+++ src/backend/routers/agent.py
@@ -96,6 +96,9 @@ async def create_agent(
     user_id = ctx.get_user_id()
     logger = ctx.get_logger()
 
+    logger.info(event="carbon id sent!", carbon_id=agent.carbon_id)
+
+
     agent_data = AgentModel(
         name=agent.name,
         description=agent.description,
